walkingkooka-tree-text/pull/133 Direction was TextDirection
- https://github.com/mP1/walkingkooka-tree-text/pull/133
- Direction was TextDirection
